Clean home and and initalize claims card from home instead
