+++ Projects/Claims/Sources/Views/ContactChatView.swift
@@ -1,4 +1,3 @@
-import Home
 import SwiftUI
 import hCore
 import hCoreUI
@@ -7,8 +6,6 @@ struct ContactChatView: View {
     let store: ClaimsStore
     let id: String
     let status: String
-
-    @EnvironmentObject var homeVm: HomeNavigationViewModel
 
     init(
         store: ClaimsStore,
