+++ Projects/Home/Project.swift
@@ -4,7 +4,7 @@ import ProjectDescriptionHelpers
 let project = Project.framework(
     name: "Home",
     targets: Set([.framework, .tests, .example]),
-    projects: ["hCore", "hCoreUI", "TravelCertificate", "TerminateContracts", "Payment", "Chat"],
+    projects: ["hCore", "hCoreUI", "TravelCertificate", "TerminateContracts", "Payment", "Chat", "Claims"],
     sdks: [],
     includesGraphQL: true
 )
