+++ Projects/Home/Sources/Components/ActiveSectionView.swift
@@ -1,4 +1,5 @@
 import Apollo
+import Claims
 import Foundation
 import PresentableStore
 import SwiftUI
@@ -6,24 +7,13 @@ import hCore
 import hCoreUI
 import hGraphQL
 
-struct ActiveSectionView<Claims: View>: View {
-    @PresentableStore var store: HomeStore
-
-    var claimsContent: Claims
-
+struct ActiveSectionView: View {
     var body: some View {
-        PresentableStoreLens(
-            HomeStore.self,
-            getter: { state in
-                state.memberContractState
-            }
-        ) { memberStateData in
-            hSection {
-                hText(L10n.HomeTab.welcomeTitleWithoutName, style: .displayXSLong)
-                    .multilineTextAlignment(.center)
-                    .fixedSize(horizontal: false, vertical: true)
-                claimsContent
-            }
+        hSection {
+            hText(L10n.HomeTab.welcomeTitleWithoutName, style: .displayXSLong)
+                .multilineTextAlignment(.center)
+                .fixedSize(horizontal: false, vertical: true)
+            ClaimsCard()
         }
     }
 }
