+++ Projects/Home/Sources/Components/TerminatedSectionView.swift
@@ -1,20 +1,19 @@
 import Apollo
+import Claims
 import Foundation
 import SwiftUI
 import hCore
 import hCoreUI
 import hGraphQL
 
-struct TerminatedSectionView<Claims: View>: View {
-    var claimsContent: Claims
-
+struct TerminatedSectionView: View {
     var body: some View {
         hSection {
             VStack(alignment: .leading, spacing: 16) {
                 hText(L10n.HomeTab.welcomeTitleWithoutName, style: .displayXSLong)
                     .multilineTextAlignment(.center)
             }
-            claimsContent
+            ClaimsCard()
         }
         .sectionContainerStyle(.transparent)
     }
